Ensure no nodes can depend on themselves even when transcoding is used (#3812)
* Factor out transcoding helpers into a private module


* Ensure node input/output validation doesn't allow transcoded self-loops


* Updated release note to avoid github warning


---------
